docker-microservices/correction/12-factors/requirements.txt
Flask==2.3.3
psycopg2-binary==2.9.7
requests==2.31.0
pytest==7.4.0